% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-nearest Neighbors Cross Validation.}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{an input data frame,}

\item{cl}{the true class value of the training data}

\item{k_nn}{an integer of the number of neighbors}

\item{k_cv}{an integer of the number of folds.}
}
\value{
a list of two objects: \code{class}; a vector of predicted classes
for each observation, and \code{cv_err}; the numeric misclassification
error.
}
\description{
k-nearest Neighbors Cross Validation.
}
\examples{
my_knn_cv(my_penguins_train, my_penguins_class, 3, 5)
my_knn_cv(my_penguins_train, my_penguins_class, 5, 5)

}
