% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My linear model function.}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{\itemize{
\item a regression formula
}}

\item{data}{\itemize{
\item a data set.
}}
}
\value{
a table with coefficient estimates, standard errors,
t-values, and t-test results for each variable.
}
\description{
This function mimics R's \code{lm()} function.
}
\examples{
my_lm(mpg ~ hp, mtcars)
my_lm(mpg ~ wt, mtcars)

}
