% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest cross-validation for penguin data set.}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{an integer of the number of folds.}
}
\value{
the numeric cross-validation error; the mean squared error
of predicted and actual penguin body mass.
}
\description{
Random Forest cross-validation for penguin data set.
}
\examples{
my_rf_cv(2)
my_rf_cv(5)

}
