% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{My T-Test Function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x, }{a numeric vector.}

\item{alternative, }{specifying the alternative hypothesis}

\item{mu, }{the null hypothesized mean.}
}
\value{
a table including the test statistic, degrees of freedom,
the alternative type, and the p-value for the test.
}
\description{
This function mimics R's \code{t.test()} function.
}
\examples{
my_t.test(rnorm(20, mean = 0, sd = 1), "two.sided", 1)
my_t.test(rnorm(20, mean = 0, sd = 1), "less", 3)

}
